word: Provide Microsoft Word Template with macros
Based on edited version of MS Word template, the twin version has
been prepared.

Note:
It is not an exact one-to-one copy. Be aware that some of provided
ribbon tools have not been used although the main of them have been
indeed. Button for clearing the formatting may reorder your text in
weird way so use it with care.